Jobbade med att försöka inkludera bilder i min befintliga CarEntity.java klass istället för att ha en separat klass för bilder. Detta visade sig vara sjukt svårt och jag lyckas skicka det till databasen men inte läsa ut det. Förtillfället så lägger man till bilder så läses dem inte ut och myCars.html crashar för att objektet blir för stort. Jag väljer att versions hantera detta iaf då jag kanske kommer tillbaka och försöker igen men lämnar detta just nu.
