--- src/main/java/com/alex/carbider/CarBider/entities/images/ImageEntity.java
@@ -1,28 +1,0 @@
-package com.alex.carbider.CarBider.entities.images;
-
-import jakarta.persistence.*;
-
-@Entity
-public class ImageEntity {
-    @Id
-    @GeneratedValue(strategy = GenerationType.AUTO)
-    private Long id;
-
-    @Lob
-    private byte[] data;
-
-    public ImageEntity(byte[] data) {
-        this.data = data;
-    }
-
-    public ImageEntity() {
-    }
-
-    public byte[] getData() {
-        return data;
-    }
-
-    public void setData(byte[] data) {
-        this.data = data;
-    }
-}
