--- src/main/java/com/alex/carbider/CarBider/entities/images/ImageRepository.java
@@ -1,8 +1,0 @@
-package com.alex.carbider.CarBider.entities.images;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface ImageRepository extends JpaRepository<ImageEntity, Long> {
-}
