+++ src/main/resources/templates/createAd.html
@@ -13,7 +13,7 @@
 
 <h1>Create an Ad!</h1>
 
-<form th:action="@{/createAd}" method="post" th:object="${carEntity}">
+<form th:action="@{/createAd}" method="post" th:object="${carEntity}" enctype="multipart/form-data">
   <label for="title">Title</label>
   <input type="text" placeholder="title..." id="title" th:field="*{Title}">
 
@@ -29,28 +29,13 @@
    <label for="email">Email</label>
    <input type="email" placeholder="email..." id="email" th:field="*{email}">
 
-
+    <label for="imageFile">Image</label>
+    <input type="file" id="imageFile" name="imageFile">
 
     <button type="submit">Post Ad!</button>
 
 </form>
 </form>
-
-<form th:action="@{/upload}" th:object="${image}" method="post" enctype="multipart/form-data">
-    <label for="image">Image</label>
-    <input type="file" placeholder="image..." id="image" th:field="*{data}" accept="image/*">
-
-    <button type="submit">Post Image</button>
-</form>
-
-
-<form action="/upload" method="post" enctype="multipart/form-data">
-    <label for="file">Choose Image:</label>
-    <input type="file" name="file" id="file" accept="image/*" required>
-    <br>
-    <button type="submit">Upload</button>
-</form>
-
 
 </body>
 </html>
